Add files via upload
The task is to write a Java program in which a user will get some attempts to guess a randomly generated number. Below are the rules of the game:
If the guessed number is bigger than the actual number, the program will respond with the message that the guessed number is higher than the actual number.
If the guessed number is smaller than the actual number, the program will respond with the message that the guessed number is lower than the actual number.
If the guessed number is equal to the actual number then the program will end with a suitable message.